requirements.txt
pillow
PySDL2
pysdl2-dll
requests
urllib3
pyinstaller